# docs/api/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CONVO API",
    "version": "1.0.0",
    "description": "The pinned contract for the implemented CONVO v1 HTTP surface."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/webhooks/meta/{appConnectionId}": {
      "get": {
        "operationId": "verifyMetaWebhook",
        "summary": "Answer the provider subscription challenge",
        "description": "Echoes `hub.challenge` when `hub.verify_token` matches the stored hash. Proving we hold the verify token is all this does: it authenticates nothing on a POST.",
        "security": [],
        "parameters": [
          {
            "name": "appConnectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "hub.mode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hub.verify_token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hub.challenge",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The challenge, echoed verbatim.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "The challenge failed. A wrong token, a wrong mode and a missing challenge are deliberately indistinguishable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "operationId": "receiveMetaWebhook",
        "summary": "Receive a signed provider delivery",
        "description": "Verifies `X-Hub-Signature-256` over the **exact raw request bytes** in constant time, refuses a replayed delivery, resolves the company from the verified asset inside the payload — never from a header, a path parameter or a query string — and journals the raw envelope before answering. A 200 means the delivery is durably journaled and nothing more.",
        "security": [],
        "parameters": [
          {
            "name": "appConnectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Hub-Signature-256",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^sha256=[0-9a-f]{64}$"
            }
          },
          {
            "name": "X-Hub-Timestamp",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "The provider envelope, verified as bytes before it is parsed.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Journaled durably.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookAccepted"
                }
              }
            }
          },
          "202": {
            "description": "Verified and understood, with nothing here to route: an asset no company has connected, or an envelope carrying no asset. Retrying cannot change that, so it is not a failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookRejected"
                }
              }
            }
          },
          "400": {
            "description": "A verified sender sent a body that will not parse. Retrying cannot help.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookRejected"
                }
              }
            }
          },
          "401": {
            "description": "The signature did not verify, or the delivery is outside the replay window.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookRejected"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "description": "The provider app secret is not configured here. Our fault, and worth retrying once it is fixed — so it is never a silent 200.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "convo_session"
      }
    },
...
    "headers": {
      "RequestId": {
        "description": "The accepted or server-generated request identifier.",
        "schema": {
          "type": "string"
        }
      },
...
    },
    "responses": {
      "Error": {
        "description": "A safe typed failure. HTTP status carries failure semantics.",
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/RequestId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "invalid_input",
                "message": "The installation bootstrap request is not valid.",
                "request_id": "request-123",
                "details": [
                  {
                    "field": "ownerEmail",
                    "code": "malformed",
                    "message": "A valid owner email is required."
                  }
                ]
              }
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "ErrorDetail": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "field",
          "code",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "request_id",
          "details"
        ],
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1
          },
          "message": {
            "type": "string",
            "minLength": 1
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorEnvelope"
          }
        }
      },
...
      "WebhookAccepted": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "status",
          "receipt_id",
          "stored",
          "duplicates",
          "request_id"
        ],
        "description": "The ACK means the delivery was journaled durably, and nothing more. Normalization, the inbox and every integration happen afterwards.",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "received"
            ]
          },
          "receipt_id": {
            "type": "string",
            "format": "uuid"
          },
          "stored": {
            "type": "integer",
            "minimum": 0
          },
          "duplicates": {
            "type": "integer",
            "minimum": 0,
            "description": "Events already journaled under the same dedupe key. Counted rather than ignored, so redelivery is visible."
          },
          "request_id": {
            "type": "string"
          }
        }
      },
      "WebhookRejected": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "status",
          "reason",
          "request_id"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "rejected"
            ]
          },
          "reason": {
            "type": "string",
            "enum": [
              "missing_header",
              "malformed_header",
              "unsupported_algorithm",
              "mismatch",
              "stale",
              "malformed_body",
              "no_asset_in_payload",
              "unknown_asset"
            ]
          },
          "request_id": {
            "type": "string"
          }
        }
      },
...
    }
  }
}